+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-11-22
 ### Added
 - `SQLiteDatabase.*` constants
+- `SQLiteDatabase.isWriteAheadLoggingEnabled` getter
 - `SQLiteDatabase#setLocale()` method
 
 [0.1.0]:  https://github.com/drydart/flutter_sqlcipher/compare/0.0.6...0.1.0